datentypen von amountFor() und thisAmount von int zu double geändert

+++ Code_Fowler/Customer.java
@@ -41,8 +41,8 @@ class Customer {
         result += "You earned " + String.valueOf(frequentRenterPoints) + " frequent renter points";
         return result;
     }
-	private int amountFor(Rental each) {
-		int thisAmount = 0;
+	private double amountFor(Rental each) {
+		double thisAmount = 0;
 		switch (each.getMovie().getPriceCode()) {
 		    case Movie.REGULAR:
 		    	thisAmount += 2;
